requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
openai
requests
websockets
streamlit
openai-agents
pymongo
motor